frontend/requirements.txt
gradio==4.13.0
requests==2.31.0
pillow==10.1.0
numpy==1.24.3